requirements.txt
fastapi[standard]
flask
boto3
humanize
python-dotenv
boto3-stubs[essential]